feat: updated color selection
Changed the tag coloring so that the list of tags does not need to match
the colored config keys in the config file.

Now, arbitrary tags can be added to each template, and still be parsed
into the Rofi selection. This makes the selection of tags easier to
expand in the future.

The script still allows for colored tags for easier differentiation.
All tags with matching color config (in config file) are printed in
color, all other default to white.