Feat/tagging (#38)
* feat: first draft adding tags to the templates

- added template.yaml as draft for potential format of payloads
- added parsing logic to select tags from file when searching
- added logic to parse payload from yaml instead of using the entire
selected file

* feat: converted first batch of scrips to yaml format

* feat: updated some more scripts after testing first batch

* fix: added missing tag for "api"

* feat: updated config to include custom colors

* feat: updated readme

* fix: fixed an issue with the '-t' command line arg

* feat: updated color selection

Changed the tag coloring so that the list of tags does not need to match
the colored config keys in the config file.

Now, arbitrary tags can be added to each template, and still be parsed
into the Rofi selection. This makes the selection of tags easier to
expand in the future.

The script still allows for colored tags for easier differentiation.
All tags with matching color config (in config file) are printed in
color, all other default to white.

* feat: added more tags

- python
- enum